wild_sugar/requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
Django==4.1
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.4
importlib-resources==6.0.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mysqlclient==2.2.0
openapi-codec==1.3.2
packaging==23.1
pkgutil_resolve_name==1.3.10
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
simplejson==3.19.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
zipp==3.16.2
